requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
kaleido>=0.2.1  # For plotly static image export
requests>=2.25.0
aiohttp>=3.7.4
python-dateutil>=2.8.2
pytz>=2023.3
cachetools>=5.3.0
psutil>=5.9.5
requests>=2.28.1
requests-futures>=1.0.0
uvicorn>=0.15.0
fastapi>=0.68.0
python-dotenv>=0.19.0
ta>=0.10.0
PyYAML>=6.0.0
aiohttp>=3.8.0
aiofiles>=0.8.0
websockets>=10.0
pybit>=5.10.0
influxdb-client>=1.48.0
jsonschema>=4.23.0
matplotlib>=3.10.1
coloredlogs>=15.0.1
discord-webhook>=1.4.1
scipy>=1.12.0
pycryptodome>=3.22.0
humanfriendly>=10.0
ccxt>=4.4.0  # Added for cryptocurrency exchange integration
GPUtil>=1.4.0
prettytable>=3.7.0  # For clean table formatting

# PDF generation and templating dependencies
weasyprint>=65.0           # For HTML to PDF conversion
pdfkit>=1.0.0              # Alternative HTML to PDF conversion (requires wkhtmltopdf)
jinja2>=3.1.0              # For HTML templating
fpdf>=1.7.2                # For PDF generation
mplfinance>=0.12.10b0      # For financial charts

# Bitcoin Beta Analysis dependencies
schedule>=1.1.0            # For automated report scheduling
reportlab>=4.0.0           # For PDF report generation

# Alpha Scanner specific
scikit-learn>=1.0.0        # For ML-based scoring
redis>=4.0.0              # Caching scan results

# Liquidation Intelligence specific
celery>=5.2.0             # Background task processing
asyncio-mqtt>=0.11.0      # Real-time message processing
sqlalchemy>=2.0.0         # Database ORM for liquidation storage
asyncpg>=0.29.0           # PostgreSQL async driver
aiosqlite>=0.19.0         # SQLite async driver
